crate: begin work on a bitcoin provider
