+++ bitcoin/src/lib.rs
@@ -8,7 +8,7 @@
 #[macro_use]
 #[doc(hidden)]
 #[cfg_attr(tarpaulin, skip)]
-pub mod prelude;
+pub mod macros;
 
 pub mod builder;
 pub mod enc;
@@ -16,7 +16,8 @@ pub mod hashes;
 pub mod nets;
 pub mod types;
 
-pub use nets::*;
+/// Common re-exports
+pub mod prelude;
 
 #[doc(hidden)]
 #[cfg(any(feature = "mainnet", feature = "testnet"))]
@@ -24,3 +25,5 @@ pub mod defaults;
 
 #[cfg(any(feature = "mainnet", feature = "testnet"))]
 pub use defaults::network::{Encoder, Network};
+
+pub use nets::*;
