+++ core/src/lib.rs
@@ -23,10 +23,6 @@ pub mod hashes;
 pub mod nets;
 pub mod ser;
 pub mod types;
+pub mod prelude;
 
-pub use builder::*;
-pub use enc::*;
-pub use hashes::*;
-pub use nets::*;
-pub use ser::*;
-pub use types::*;
+pub use prelude::*;
