+++ core/src/prelude.rs
@@ -1,0 +1,8 @@
+//! Re-exports of common traits.
+pub use crate::{
+    builder::TxBuilder,
+    enc::AddressEncoder,
+    nets::Network,
+    ser::ByteFormat,
+    types::Transaction,
+};
