+++ ledger/Cargo.toml
@@ -12,27 +12,30 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 thiserror = "1.0.10"
-byteorder = "1.3.1"
-matches = "0.1.8"
-lazy_static = "1.3.0"
-blake2b_simd = "0.5.10"
 serde = { version = "1.0", features = ["derive"] }
-cfg-if = "0.1.7"
 futures = "0.3.5"
 async-trait = "0.1.35"
 
+# native
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+lazy_static = "1.3.0"
+byteorder = "1.3.1"
+blake2b_simd = "0.5.10"
 libc = "0.2.50"
+cfg-if = "0.1.7"
+matches = "0.1.8"
 
+# native
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies.hidapi]
 version = "1.1.1"
 default-features = false
 features=["linux-static-hidraw"]
 
+# linux native only
 [target.'cfg(target_os = "linux")'.dependencies]
 nix = "0.13.0"
 
-# For wasm compiling
+# WASM
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 wasm-bindgen = "0.2.60"
 wasm-bindgen-futures = "0.4.10"
@@ -42,6 +45,5 @@ js-sys = "0.3.37"
 serial_test = "0.3.2"
 
 [features]
-default = []
 browser = []
 node = []
