+++ provider/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "provider"
+version = "0.1.0"
+authors = ["James Prestwich <james@prestwi.ch>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+async-trait = "0.1.36"
+
+futures-core = { version = "0.3.5", default-features = false }
+futures-util = { version = "0.3.5", default-features = false }
+futures-timer = "3.0.2"
+pin-project = { version = "0.4.20", default-features = false }
+
+riemann-core = { path = "../core" }
+rmn-btc = { path = "../bitcoin" }
