+++ psbt/src/lib.rs
@@ -33,7 +33,7 @@ use rmn_bip32::{
     KeyFingerprint, XPub,
 };
 
-use riemann_core::{builder::TxBuilder, enc::AddressEncoder, ser::ByteFormat, tx::Transaction};
+use riemann_core::prelude::*;
 
 use rmn_btc::{
     builder::LegacyBuilder,
